+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-03-25 - Jorge
+### Added
+- **Ambient lights** now affects every figure's color. The color modification gets applied when checking which is the closest point for each ray.
 ## 2022-03-22 - Jorge
 ### Added
 - **change_color_matrix** now stores the color of the figure touched. There's a problem with collisions involving spheres (the transition isn't smooth).